cob: Add support for media embeds in COBs
This change allows for files (blobs) to be embedded into COB entries.
This allows for things like image attachments in issue comments for example.

For now, we only enable this in the `issue` COB.

The way it works is that relevant COB actions carry metadata about which
files are attached to them; and we store those files as blobs inside the
COB entry's *tree* object, under an `embeds/` folder.

The `Embed<T>` type is used for the above, and either carries actual
content, or carries a content-id. Retrieving the actual content is as
simple as asking the repository for a blob with that content-id.

As a possible future extension, MIME types could be stored alongside the
files in a "metadata" file. This could help clients display the content
appropriately.

+++ radicle-cob/src/object/collaboration/info.rs
@@ -17,7 +17,7 @@ pub struct ChangeGraphInfo {
     /// The ID of the object
     pub object_id: ObjectId,
     /// The number of nodes in the change graph of the object
-    pub number_of_nodes: u64,
+    pub number_of_nodes: usize,
     /// The "tips" of the change graph, i.e the object IDs pointed to by
     /// references to the object
     pub tips: BTreeSet<Oid>,
